docker/requirements.txt
reportlab
qrcode
python-barcode
pillow
boto3